Add NewTestEtcd test utility function.
* Refactor store test helper to use this.
* Refactor etcd test function.
* Refactor queue test functions.


Removed merge artifact from changelog
